+++ web/src/server/domain/layout/repo/layoutLocalFileRepository.ts
@@ -3,8 +3,8 @@ import { REPO_LAYOUT_FILE } from "~/lib/basic/const";
 import { codes } from "~/lib/error/codes";
 import AppError from "~/lib/error/error";
 import { type Reader } from "../../../driver/Reader/Reader";
-import { parseAdjustedWidgetLayout } from "../../config/services/parseWidgetConfig.service";
 import type { AdjustedWidgetLayout } from "../adjustedWidgetLayout";
+import { parseAdjustedWidgetLayout } from "../services/parse.service";
 import type { LayoutRepository } from "./layoutRepository";
 
 export class LayoutLocalFileRepository implements LayoutRepository {
